+++ src/SkolplattformenElevApi/Models/Internal/Calendar/CalendarResponse.cs
@@ -3,7 +3,7 @@
 namespace SkolplattformenElevApi.Models.Internal.Calendar;
 
 // Root myDeserializedClass = JsonSerializer.Deserialize<Root>(myJsonResponse);
-internal class CalendarItem
+internal class CalendarItemData
 {
     [JsonPropertyName("id")]
     public string Id { get; set; }
@@ -36,7 +36,7 @@ internal class CalendarResponse
     public object Error { get; set; }
 
     [JsonPropertyName("Data")]
-    public List<CalendarItem>? Data { get; set; }
+    public List<CalendarItemData>? Data { get; set; }
 }
