+++ src/SkolplattformenElevApi/Models/School.cs
@@ -1,0 +1,11 @@
+﻿namespace SkolplattformenElevApi.Models;
+
+public class School
+{
+    public Guid Id { get; set; }
+    public Guid ExternalId { get; set; }
+    public string Name { get; set; }
+    public string Url { get; set; }
+        
+        
+}
