+++ src/SkolplattformenElevApi/Models/Teacher.cs
@@ -1,0 +1,10 @@
+﻿namespace SkolplattformenElevApi.Models;
+
+public class Teacher
+{
+    public int Id { get; set; }
+    public string Firstname { get; set; }
+    public string Lastname { get; set; }
+    //  public string Phone { get; set; }
+    public string Email { get; set; }
+}
